+++ game_patch/misc/alpine_settings.cpp
@@ -529,10 +529,6 @@ bool alpine_player_settings_load(rf::Player* player)
         set_big_hud(g_alpine_game_config.big_hud);
         processed_keys.insert("BigHUD");
     }
-    if (settings.count("ReticleScale")) {
-        g_alpine_game_config.set_reticle_scale(std::stof(settings["ReticleScale"]));
-        processed_keys.insert("ReticleScale");
-    }
     if (settings.count("SniperScopeColor")) {
         auto color_override = parse_hex_color_string(settings["SniperScopeColor"]);
         if (color_override) {
@@ -632,6 +628,16 @@ bool alpine_player_settings_load(rf::Player* player)
             xlog::warn("Invalid reticle color override: {}", settings["ReticleLockedColor"]);
         }
         processed_keys.insert("ReticleLockedColor");
+    }
+    if (settings.count("ReticleScale")) {
+        const float scale = std::stof(settings["ReticleScale"]);
+        if (scale == 1.0f) {
+            g_alpine_game_config.clear_reticle_scale();
+        }
+        else {
+            g_alpine_game_config.set_reticle_scale(scale);
+        }
+        processed_keys.insert("ReticleScale");
     }
     if (settings.count("DamageNotifyTextScale")) {
         const float scale = std::stof(settings["DamageNotifyTextScale"]);
@@ -1112,7 +1118,6 @@ void alpine_player_settings_save(rf::Player* player)
     // UI
     file << "\n[UISettings]\n";
     file << "BigHUD=" << g_alpine_game_config.big_hud << "\n";
-    file << "ReticleScale=" << g_alpine_game_config.reticle_scale << "\n";
     if (g_alpine_game_config.sniper_scope_color_override) {
         file << "SniperScopeColor=" << format_hex_color_string(*g_alpine_game_config.sniper_scope_color_override) << "\n";
     }
@@ -1142,6 +1147,9 @@ void alpine_player_settings_save(rf::Player* player)
     }
     if (g_alpine_game_config.teammate_label_color_override) {
         file << "TeammateLabelColor=" << format_hex_color_string(*g_alpine_game_config.teammate_label_color_override) << "\n";
+    }
+    if (g_alpine_game_config.reticle_scale) {
+        file << "ReticleScale=" << *g_alpine_game_config.reticle_scale << "\n";
     }
     if (g_alpine_game_config.world_hud_damage_text_scale) {
         file << "DamageNotifyTextScale=" << *g_alpine_game_config.world_hud_damage_text_scale << "\n";