+++ game_patch/misc/alpine_settings.h
@@ -45,12 +45,6 @@ struct AlpineGameSettings
         scanner_sensitivity_modifier = std::clamp(mod, min_sens_mod, max_sens_mod);
     }
 
-    float reticle_scale = 1.0f;
-    void set_reticle_scale(float scale)
-    {
-        reticle_scale = std::clamp(scale, 0.0f, 100.0f);
-    }
-
     float level_sound_volume = 1.0f;
     void set_level_sound_volume(float scale)
     {
@@ -165,6 +159,20 @@ struct AlpineGameSettings
     std::optional<uint32_t> reticle_locked_color_override{};
 
     // hud scale overrides
+    std::optional<float> reticle_scale{};
+    float get_reticle_scale() const
+    {
+        return reticle_scale.value_or(1.0f);
+    }
+    void set_reticle_scale(float scale)
+    {
+        reticle_scale = std::clamp(scale, 0.0f, 100.0f);
+    }
+    void clear_reticle_scale()
+    {
+        reticle_scale.reset();
+    }
+
     std::optional<float> world_hud_damage_text_scale{};
     float get_world_hud_damage_text_scale() const
     {